Move the pypi listener into a dedicated library
This makes testing easier and cleans up the interface by pushing some
initialization into a dedicated NewListener function.
