+++ tools/pypi_rss/listener/listener_test.go
@@ -1,4 +1,4 @@
-package main
+package listener
 
 import (
 	"context"
